update results to differentiate between min compatibility and optional (#133)
At a minimum subgraphs should support `_service { sdl }`, single `@key` and `@link` directives. This allows for API-side joins and use of entities in other subgraphs. Those checks will continue to be marked as :green_square:/:red_square:.

Additional functionality listed below is required for full Federation support. If functionality is unsupported, it will be marked as :black_square_button: in the compatibility matrix.

* multiple/complex `@key`s
* using `@requires` to provide additional information for computed fields
* path denormalization using `@provides`
* ability to relax single source of truth for entity fields using `@shareable`
* using `@override` to easily migrate fields between subgraphs
* hiding fields by using `@inaccessible`
* using `@tag` for contracts
* tracing information using `ftv1`