update @requires test to verify the functionality
Updating `users` and `products` subgraph schemas to allow for testing the support of `@requires` functionality in the implementing subgraph. This is a breaking change as we modify the schema so all the existing implementations will be impacted.

Changes:
* `users` subgraph
```graphql
type User @key(fields:"email") {
  email:ID!
  name: String
  totalProductsCreated: Int
  yearsOfEmployment: Int! # <-- new field
}
```
* `products` subgraph
```graphql
extend type User @key(fields: "email") {
  averageProductsCreatedPerYear: Int @requires(fields: "yearsOfEmployment") # <-- test `@requires` functionality
  email: ID! @external
  name: String @override(from: "users")
  totalProductsCreated: Int @external
  yearsOfEmployment: Int! @external # <-- field from the users subgraph used by @requires
}
```

Related Issues:

* https://github.com/apollographql/apollo-federation-subgraph-compatibility/issues/128