fix schema comparison tests (#152)
Schema comparison was incorrectly comparing implemented schema against the expected schema only making basic comparisons and only checking for fields on the implemented schema.

Issues:
* If expected schema defined more elements or types they were never compared.
* If fields were non-nullable then tests were not comparing actual return types.
* Arguments were not compared

Schema comparison will now perform the following checks and return error summary if any of them fail:
* implementing type has to declare same amount of fields as expected one
* implementing type declares all the expected fields
* compares nullability of fields
* compares return type of fields
* compares arguments (nullability and type)

NOTE: Schema comparison does not currently work with lists as we don't have any list types defined in the schema.
NOTE2: Sadly Jest does not allow specifying custom messages to the assertions and does not support nested tests. `jest-expect-message` is unmaintained and not compatible with latest versions of Jest. Creating custom matcher seemed like an overkill so I made a workaround of capturing error messages into a string and expecting it to be empty.